# openapi-swagger.yaml
openapi: 3.0.3
info:
  title: KoZnaZna Platform API
  description: |
    Kompletna API dokumentacija za KoZnaZna Platform - mikroservisnu arhitekturu za upravljanje pub kviz platformom.
    
    Sistem se sastoji od tri glavna servisa:
    - **API Gateway** (Port 8000) - Centralna tačka za autentifikaciju i rutiranje
    - **Organization Service** (Port 8001) - Upravljanje organizacijama i članovima
    - **Quiz Service** (Port 8002) - Upravljanje kvizovima, timovima i ligama
    
    Kreirana za Master rad 2025.
  version: 1.0.0
  contact:
    name: KoZnaZna Platform
    email: contact@koznazna.com
servers:
  - url: http://localhost:8000/api
    description: API Gateway (Development)
  - url: http://localhost:8001/api
    description: Organization Service (Internal)
  - url: http://localhost:8002/api
    description: Quiz Service (Internal)
security:
  - BearerAuth: []
  - InternalAuth: []
paths:
...
  /leagues:
    get:
      tags:
        - Leagues
      summary: Lista liga
      description: Dobijanje liste svih liga (javno dostupno)
      security: []
      responses:
        '200':
          description: Lista liga
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/League'
    post:
      tags:
        - Leagues
      summary: Kreiranje lige
      description: Kreiranje nove lige (SUPER_ADMIN/ORG_ADMIN/ORG_MEMBER)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeagueCreate'
      responses:
        '201':
          description: Uspešno kreirana liga
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
  /leagues/{id}:
    get:
      tags:
        - Leagues
      summary: Detalji lige
      description: Dobijanje detalja određene lige (javno dostupno)
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalji lige
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
    put:
      tags:
        - Leagues
      summary: Ažuriranje lige
      description: Ažuriranje podataka lige (SUPER_ADMIN/ORG_ADMIN/ORG_MEMBER)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeagueCreate'
      responses:
        '200':
          description: Uspešno ažurirana liga
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
    delete:
      tags:
        - Leagues
      summary: Brisanje lige
      description: Brisanje lige iz sistema (SUPER_ADMIN/ORG_ADMIN/ORG_MEMBER)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Uspešno obrisana liga
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Laravel Sanctum token za autentifikaciju frontend klijenata
    InternalAuth:
      type: apiKey
      in: header
      name: X-Internal-Auth
      description: Interni token za komunikaciju između servisa
  schemas:
...
    League:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Zimska Liga 2025"
        description:
          type: string
          nullable: true
          example: "Zimska sezona kviz lige"
        organization_id:
          type: integer
          example: 1
        start_date:
          type: string
          format: date
          example: "2025-01-01"
        end_date:
          type: string
          format: date
          nullable: true
          example: "2025-03-31"
        is_active:
          type: boolean
          example: true
        total_rounds:
          type: integer
          example: 10
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LeagueCreate:
      type: object
      required:
        - name
        - organization_id
        - start_date
        - total_rounds
      properties:
        name:
          type: string
          example: "Zimska Liga 2025"
        description:
          type: string
          nullable: true
          example: "Zimska sezona kviz lige"
        organization_id:
          type: integer
          example: 1
        start_date:
          type: string
          format: date
          example: "2025-01-01"
        end_date:
          type: string
          format: date
          nullable: true
          example: "2025-03-31"
        total_rounds:
          type: integer
          minimum: 1
          example: 10
        is_active:
          type: boolean
          default: true
...